Refactor script

+++ api/v1/views/users.py
@@ -1,30 +1,15 @@
 #!/usr/bin/python3
 '''Contains the users view for the API.'''
 from flask import jsonify, request
-from werkzeug.exceptions import NotFound, MethodNotAllowed, BadRequest
+from werkzeug.exceptions import NotFound, BadRequest
 
 from api.v1.views import app_views
 from models import storage
 from models.user import User
 
 
-@app_views.route('/users', methods=['GET', 'POST'])
-@app_views.route('/users/<user_id>', methods=['GET', 'DELETE', 'PUT'])
-def handle_users(user_id=None):
-    '''The method handler for the users endpoint.
-    '''
-    handlers = {
-        'GET': get_users,
-        'DELETE': remove_user,
-        'POST': add_user,
-        'PUT': update_user,
-    }
-    if request.method in handlers:
-        return handlers[request.method](user_id)
-    else:
-        raise MethodNotAllowed(list(handlers.keys()))
-
-
+@app_views.route('/users', methods=['GET'])
+@app_views.route('/users/<user_id>', methods=['GET'])
 def get_users(user_id=None):
     '''Gets the user with the given id or all users.
     '''
@@ -38,18 +23,21 @@ def get_users(user_id=None):
     return jsonify(users)
 
 
-def remove_user(user_id=None):
+@app_views.route('/users/<user_id>', methods=['DELETE'])
+def remove_user(user_id):
     '''Removes a user with the given id.
     '''
-    user = storage.get(User, user_id)
-    if user:
-        storage.delete(user)
-        storage.save()
-        return jsonify({}), 200
+    if user_id:
+        user = storage.get(User, user_id)
+        if user:
+            storage.delete(user)
+            storage.save()
+            return jsonify({}), 200
     raise NotFound()
 
 
-def add_user(user_id=None):
+@app_views.route('/users', methods=['POST'])
+def add_user():
     '''Adds a new user.
     '''
     data = request.get_json()
@@ -64,6 +52,7 @@ def add_user(user_id=None):
     return jsonify(new_user.to_dict()), 201
 
 
+@app_views.route('/users/<user_id>', methods=['PUT'])
 def update_user(user_id=None):
     '''Updates the user with the given id.
     '''
